requirements.txt
python-dotenv
httpx
azure-identity
uvicorn
starlette
sse-starlette
fastapi
azure-ai-projects
azure-ai-agents
a2a-sdk